+++ app/src/main/java/com/innovatrics/dot/samples/nfcreading/NfcTravelDocumentReaderFactory.kt
@@ -1,23 +1,8 @@
 package com.innovatrics.dot.samples.nfcreading
 
-import android.content.res.Resources
 import com.innovatrics.dot.nfc.NfcTravelDocumentReader
-import com.innovatrics.dot.nfc.NfcTravelDocumentReaderConfiguration
-import com.innovatrics.dot.nfc.NfcTravelDocumentReaderFactory
-import com.innovatrics.dot.nfc.security.CertificatesFactory
-import com.innovatrics.dot.samples.R
-import java.security.cert.X509Certificate
 
-fun createNfcTravelDocumentReader(resources: Resources): NfcTravelDocumentReader {
-    val authorityCertificates = createAuthorityCertificates(resources)
-    val configuration = NfcTravelDocumentReaderConfiguration.Builder()
-        .authorityCertificates(authorityCertificates)
-        .build()
-    return NfcTravelDocumentReaderFactory.create(configuration)
-}
+interface NfcTravelDocumentReaderFactory {
 
-fun createAuthorityCertificates(resources: Resources): Set<X509Certificate> {
-    resources.openRawResource(R.raw.master_list).use { inputStream ->
-        return CertificatesFactory.create(inputStream)
-    }
+    fun create(): NfcTravelDocumentReader
 }