+++ app/src/main/java/com/innovatrics/dot/samples/ui/ByteArraySerializer.kt
@@ -1,7 +1,7 @@
 package com.innovatrics.dot.samples.ui
 
 import com.google.gson.JsonElement
-import com.google.gson.JsonObject
+import com.google.gson.JsonPrimitive
 import com.google.gson.JsonSerializationContext
 import com.google.gson.JsonSerializer
 import java.lang.reflect.Type
@@ -9,8 +9,6 @@ import java.lang.reflect.Type
 class ByteArraySerializer : JsonSerializer<ByteArray> {
 
     override fun serialize(src: ByteArray?, typeOfSrc: Type?, context: JsonSerializationContext?): JsonElement {
-        return JsonObject().apply {
-            addProperty("size", src?.size)
-        }
+        return JsonPrimitive("${src?.size} bytes")
     }
 }